requirements.txt
streamlit>=1.0
streamlit-extras>=0.1.5
google-generativeai>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
Pillow>=9.0.0
